added animations

+++ components/FavoritesComponent.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import * as Animatable from 'react-native-animatable';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -29,19 +30,38 @@ class Favorites extends Component {
         const renderFavoriteItem = ({item}) => {
             const rightButton = [
                 {
-                    text: 'Delete',
+                    text: 'Delete', 
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite campsite ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
             return (
                 <Swipeout right={rightButton} autoClose={true}>
+                <Animatable.View animation='fadeInRightBig' duration={2000}>
                 <ListItem
                     title={item.name}
                     subtitle={item.description}
                     leftAvatar={{source: {uri: baseUrl + item.image}}}
                     onPress={() => navigate('CampsiteInfo', {campsiteId: item.id})}
                 />
+                </Animatable.View>
                 </Swipeout>
             );
         };
